Check Cache vs Current

Checks for cache file, makes one if it doesn't exist, updates if it's out of date, and loads it otherwise. Returns a list with the cached values. Option to skip overwrite of cache and just return lineages. Custom target files for updateHierarchies() can be supplied using aliases_url and lineages_url

Usage

checkCache(overwrite = TRUE, network = FALSE, ...)

Arguments

overwrite

set to false to compare cache to file and return data without writing

...

Pass alternate lineages_url and aliases_url to updateHierarchies

output

logical, TRUE for network (produced by fromDFtoNetwork) FALSE for table

Value

depending on the value of output

table of lineage hierarchies with fixed aliases

a Node class network object describing the hierarchy of lineages