Check Cache vs Current

Checks for cache file, makes one if it doesn't exist, updates if it's out of date, and loads it otherwise. Returns a list with the cached values. Option to skip overwrite of cache and just return lineages. Custom target files for updateHierarchies() can be supplied using aliases_url and lineages_url

checkCache(overwrite = TRUE, ...)

Arguments

overwrite

set to false to compare cache to file and return data

...

Pass alternate lineages_url and aliases_url to updateHierarchies

Value

a list with three or more elements

lineages list with fixed aliases

hierarchy list

a Node class network object describing the hierarchy of lineages

a table of so-called orphan lineages that call on non-existent parents

file.info output showing most recent update time for cache

new lineages between cache and update

new aliases between cache and update